This commit introduces the new data format with GUIDs into Atlas. The code has been modified slightly to read this new format, and the Atlas-enUS file has been outfitted with it. So far, work has started on the BG module, but no other modules yet. So: NO MODULES WILL WORK WITH THIS VERSION--DISABLE THEM--only the core Atlas addon will work! I'll be updating the rest of the modules very soon.
I've tried hard to make sure the new Atlas-enUS is UP TO DATE with all the recent changes. This required some manual syncing (using revision tools of course), but it's still possible that something slipped by, so please give it a look and let me know if anything is busted or incorrect.

The main reason for this update is for the Atlasmod.com website, which will now be using the exact same file to generate the individual map web pages! If you have any comments or questions about this, let me know.
In addition, due to request, I've changed the frame strata back to MEDIUM so the window can interact with the other windows--it can be raised over others with a click, and others can be raised over it with a click. As a result this reintroduced the problem of the bottom and side bars overlapping the window, so I added a little fix for them--just changed their frame strata when Atlas loads. I don't know if this will introduce any strange behavior elsewhere, but it seems to work fine for the time being--let me know if problems emerge.

Also made some other small changes, and updated the change log.

- changed the Atlas data format to include GUID information (no in-game change)
- changed the frame strata back to MEDIUM so Atlas can cover/be-covered-by other windows
- added a fix for some UI elements that appear above the Atlas window
- added a mark (1) for Patchwerk to the map of Naxx
- changed the entry colors for the Abomination Wing in Naxx